refactor: 메뉴 관련 클래스를 메뉴 패키지 아래에 추가.
