+++ src/main/java/kitchenpos/menu/repository/JpaMenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.repository;
 
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
