+++ src/main/java/kitchenpos/menu/repository/JpaMenuRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.repository;
 
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
