+++ src/main/java/kitchenpos/menu/repository/MenuGroupRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.repository;
+
+import kitchenpos.menu.domain.MenuGroup;
 
 import java.util.List;
 import java.util.Optional;
