+++ src/main/java/kitchenpos/menu/repository/MenuRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.repository;
+
+import kitchenpos.menu.domain.Menu;
 
 import java.util.List;
 import java.util.Optional;
