+++ src/main/java/kitchenpos/menu/service/MenuGroupService.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.menu.service;
 
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
